Make report builder and homepage builder settings the default
